requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.5.0=py39haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py39haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
ca-certificates=2023.08.22=haa95532_0
category-encoders=2.6.3=pypi_0
cffi=1.15.1=py39h2bbff1b_3
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=py39haa95532_0
colorlover=0.3.0=pypi_0
comm=0.2.0=pypi_0
contourpy=1.0.5=py39h59b6b97_0
cufflinks=0.17.3=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
dask=2023.12.0=pypi_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2023.12.0=pypi_0
eli5=0.13.0=pypi_0
entrypoints=0.4=py39haa95532_0
exceptiongroup=1.0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
feather-format=0.4.1=pypi_0
featurewiz=0.4.7=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2023.12.1=pypi_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
importlib-metadata=7.0.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.25.0=py39h9909e9c_0
ipython=8.15.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.1.1=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.1.2=py39haa95532_0
joblib=1.2.0=py39haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py39haa95532_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter_client=7.4.9=py39haa95532_0
jupyter_core=5.3.0=py39haa95532_0
jupyter_server=1.23.4=py39haa95532_0
jupyterlab-widgets=3.0.9=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.4.4=py39hd77b12b_0
krb5=1.20.1=h5b6d351_0
lazytransform=1.10=pypi_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lightgbm=4.1.0=pypi_0
locket=1.0.0=pypi_0
lxml=4.9.3=py39h09808a7_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.7.2=py39haa95532_0
matplotlib-base=3.7.2=py39h4ed8f06_0
matplotlib-inline=0.1.6=py39haa95532_0
mistune=0.8.4=py39h2bbff1b_1000
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
msgpack=1.0.7=pypi_0
munkres=1.1.4=py_0
nbclassic=0.5.5=py39haa95532_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.5.4=py39haa95532_0
nbformat=5.9.2=py39haa95532_0
nest-asyncio=1.5.6=py39haa95532_0
networkx=3.2.1=pypi_0
notebook=6.5.4=py39haa95532_1
notebook-shim=0.2.2=py39haa95532_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.0=py39h055cbcc_0
numpy-base=1.26.0=py39h65a83cf_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
packaging=23.1=py39haa95532_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.4.1=pypi_0
patsy=0.5.3=py39haa95532_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py39h045eedc_0
pip=23.2.1=py39haa95532_0
platformdirs=3.10.0=py39haa95532_0
plotly=5.9.0=py39haa95532_0
ply=3.11=py39haa95532_0
prometheus_client=0.14.1=py39haa95532_0
prompt-toolkit=3.0.36=py39haa95532_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=14.0.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py39haa95532_1
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.7=py39hd77b12b_0
pyqt5-sip=12.11.0=py39hd77b12b_0
pyrsistent=0.18.0=py39h196d8e1_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39haa95532_0
python-graphviz=0.20.1=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py39haa95532_0
pywin32=305=py39h2bbff1b_0
pywinpty=2.0.10=py39h5da7b33_0
pyyaml=6.0.1=pypi_0
pyzmq=23.2.0=py39hd77b12b_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
qtwebkit=5.212=h2bbfb41_5
scikit-learn=1.2.2=pypi_0
scipy=1.11.3=py39h309d312_0
seaborn=0.12.2=py39haa95532_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=68.0.0=py39haa95532_0
sip=6.6.2=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39haa95532_1
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=py39haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.14.0=py39h9128911_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
tblib=3.0.0=pypi_0
tenacity=8.2.2=py39haa95532_0
terminado=0.17.1=py39haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py39haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
tornado=6.3.2=py39h2bbff1b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py39haa95532_0
typing-extensions=4.7.1=py39haa95532_0
typing_extensions=4.7.1=py39haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=0.58.0=py39haa95532_4
wheel=0.41.2=py39haa95532_0
widgetsnbextension=4.0.9=pypi_0
winpty=0.4.3=4
xgboost=1.6.2=pypi_0
xlrd=2.0.1=pypi_0
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zict=3.0.0=pypi_0
zipp=3.11.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
